Avoid warnings when trying to use doubles as arguments for an int parameter.
git-svn-id: https://svn.dealii.org/trunk@3381 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/deal.II/include/numerics/time_dependent.h
@@ -1707,7 +1707,7 @@ struct TimeStepBase_Tria<dim>::RefinementFlags
 				      * Exception
 				      */
     DeclException1 (ExcInvalidValue,
-		    int,
+		    double,
 		    << "The following value does not fulfil the requirements: " << arg1);
 };
